vspfilter: Get the parameters of a video stream from a GstVideoFilter instead of a GstVideoFrame structure
A GstVideoFrame structure is obtained by mapping a GstBuffer, using
a GstAllocator. To handle dmabuf descriptors without the mapping,
video parameters, which are a video width, height and format respectively,
have to be obtained from other than the GstVideoFrame.

However only the stride of an output GstVideoFrame has been used yet
because the stride from a GstVideoFilter is simply calculated by
multiplying a width by a pixel stride.
This is going to be fixed later.

This is preliminary for the dmabuf handling support.