vspfilter: Override the set_caps method of the gstbasetransform class
This is preliminary for removing the gst_vsp_filter_transform_frame
method, which is never called.

The set_caps method of the videofilter class, which is the parent
class of the vspfilter, sets the in-place conversion mode
if the inheritance class does not register the transform_frame method.
This patch avoids the unnecessary mode change to the in-place mode
when removing the transform_frame method.